refactor(diagram): type 파일 분리, 일부 주석 추가
